Fixed bug when parsing link geometries
Fixed a bug that occurs when a link in the URDF contains both visual- and collision-elements.
Function calls to NextSiblingElement() were missing an optional argument (i.e. "visual" and "collision"). As a result the function did return any xml-sibling. This caused collision-elements to be additionally loaded as visual-elements and vice versa (depending on their order in the URDF).